QueueUpdater second version, added JobLog class
Initial Track QueueItems job created, not tested yet

JobLog class added, purpose is to log the las time a job was run successfully,
as well as any failures. These could be stored in DB as well, but would require
a dedicated DB table, which seems overkill. The last success information is
required to determine whether a status record is reliable or not (say if SOAP
doesn't return an answer for an hour, for example).

QueueUpdater can now handle all normal situations and a few borderline cases.
It cannot correctly handle a situation where the time_in_queue attribute returned
by SOAP is not accurate (for example because of lag).