GetTeamContactsJob now fetches all contacts with 1 query
GetTeamContacts added to Backendservice
API changed for BackendService for GetAgentContacts(currently unused)

Contact model updated, information for service_id and contact type added. Phone number removed as unused.
Contact is now aware of the team it belongs to via service.

GetTeamContactsJob will attempt to find the agent it belongs to, and if none is found, assumes the agent no longer exists(quit/removed from system for other reason). Such contacts are not stored
It will also attempt to find the service it belongs to, and if one with a matching name is not found, nil is assigned to service_id.